refactor namespace of a utility class
