+++ CxRestClient/CxMnoPolicies.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Net.Http;
 using System.Text;
 using System.Threading;
+using CxRestClient.Utility;
 
 namespace CxRestClient
 {
