+++ CxRestClient/CxMnoPolicyRules.cs
@@ -5,8 +5,8 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Net.Http;
-using System.Text;
 using System.Threading;
+using CxRestClient.Utility;
 
 namespace CxRestClient
 {
