+++ CxRestClient/CxSastGenerateScanReport.cs
@@ -9,6 +9,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using CxRestClient.Utility;
 
 namespace CxRestClient
 {
